@@ -10,7 +10,7 @@
     "gen:styles:light": "smui-theme compile src/renderer/css/smui.css -i src/sass/light",
     "gen:styles:dark": "smui-theme compile src/renderer/css/smui-dark.css -i src/sass/dark",
     "gen:styles:global": "sass src/sass/global.scss src/renderer/css/global.css",
-    "gen:styles": "npm run gen:styles:light && npm run gen:styles:dark && npm run gen:styles:global",
+    "gen:styles": "npm run gen:styles:global",
     "gen:grammars": "nearleyc src/nearley/data.ne -o src/main/lib/grammars/dataGrammar.mjs",
     "gen:data": "set -a; source .tools.env; set +a; vite-node -c vite.tools.config.ts tools/generateAllData.ts",
     "gen": "npm run gen:styles",
@@ -29,32 +29,22 @@
     "@floating-ui/dom": "^1.6.3",
     "@html-eslint/eslint-plugin": "^0.21.0",
     "@html-eslint/parser": "^0.21.0",
-    "@material/dom": "^14.0.0",
-    "@smui/button": "^7.0.0-beta.15",
-    "@smui/chips": "^7.0.0-beta.16",
-    "@smui/circular-progress": "^7.0.0-beta.15",
-    "@smui/common": "^7.0.0-beta.15",
-    "@smui/dialog": "^7.0.0-beta.15",
-    "@smui/paper": "^7.0.0-beta.15",
-    "@smui/textfield": "^7.0.0-beta.15",
-    "@smui/tooltip": "^7.0.0-beta.16",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
-    "@tsconfig/svelte": "^5.0.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@tsconfig/svelte": "^5.0.4",
     "@typescript-eslint/eslint-plugin": "^6.13.1",
     "@typescript-eslint/parser": "^6.13.1",
     "electron": "27.1.2",
     "electron-affinity": "^1.2.4",
     "eslint": "^8.54.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-svelte": "^2.35.1",
+    "eslint-plugin-svelte": "^2.46.0",
     "material-icons": "^1.13.12",
     "nearley": "^2.20.1",
     "object-hash": "^3.0.0",
     "sass": "^1.69.5",
     "seedrandom": "^3.0.5",
-    "smui-theme": "^7.0.0-beta.15",
     "source-map-support": "^0.5.21",
-    "svelte": "^4.2.7",
+    "svelte": "^5.1.16",
     "vite-node": "^1.6.0",
     "vite-tsconfig-paths": "^4.2.1",
     "yaml": "^2.3.4"
