@@ -1,65 +1,131 @@
-<Dialog bind:open={open}>
-  {#if isNotNullish(title)}
-    <Title>{title}</Title>
-  {/if}
-  
-  <Content>
-    {#if isNotNullish(message)}
-      {message}
+<!-- svelte-ignore a11y_no_static_element_interactions -->
+<!-- svelte-ignore a11y_click_events_have_key_events -->
+<div
+  style:display={open ? "block" : "none"}
+  style:position="fixed"
+  style:inset="0"
+  style:width="100%"
+  style:height="100%"
+  style:background-color="#00000055"
+  style:backdrop-filter="blur(10px)"
+  style:z-index={zIndexes.dialogLayer}
+  onclick={cancelButtonClicked}
+>
+</div>
+<div
+  style:position="absolute"
+  style:margin="auto"
+  style:background-color={colors.primarySurface}
+  style:box-shadow="5px 5px 5px #00000070"
+  style:border-radius="20px"
+  style:top="50%"
+  style:left="50%"
+  style:transform="translate(-50%, -50%)"
+  style:z-index={zIndexes.dialogLayer}
+>
+  <Stack
+    alignment="fill"
+    direction="vertical"
+    distribution="fill"
+    maxHeight="100vh"
+    minSpacing={20}
+    padding={50}
+  >
+    {#if isNotNullish(title)}
+      <div use:textStyle={"title"}>
+        {title}
+      </div>
     {/if}
-    
+  
+    {#if isNotNullish(message)}
+      <div
+        style:min-height="20px"
+        style:overflow="scroll"
+      >
+        <Stack
+          alignment="start"
+          direction="vertical"
+          distribution="fill"
+          minSpacing={10}
+        >
+          {#each message.split("\n") as paragraph}
+            <div use:textStyle={"content"}>{paragraph}</div>
+          {/each}
+        </Stack>
+      </div>
+    {/if}
+  
     {#if isNotNullish(inputInfo)}
-      <Textfield
-        bind:this={textfield}
-        input$accept={inputInfo.fileExtension}
-        label={inputInfo.title}
-        type={inputInfo.type}
-        bind:value={textfieldValue}
+      <FileInput
+        allowedFileTypes={inputInfo.fileExtension}
+        title={inputInfo.title}
+        bind:value={selectedFileName}
         bind:files={files}
       />
     {/if}
-  </Content>
-  
-  <Actions>
-    {#if hasCancelButton}
-      <Button on:click={onCancel}>
-        <Label>Cancel</Label>
-      </Button>
-    {/if}
-    
-    <Button
-      bind:this={submitButton}
-      variant="raised"
-      on:click={submitButtonClicked}
+    <Stack
+      alignment="fill"
+      direction="horizontal"
+      distribution="fill"
+      minSpacing={20}
     >
-      <Label>{submitButtonLabel}</Label>
-    </Button>
-  </Actions>
-</Dialog>
+      {#if hasCancelButton}
+        <Button
+          style="fill"
+          flexGrow={true}
+          isDestructive={true}
+          onClick={cancelButtonClicked}
+          title="Cancel"
+        />
+      {/if}
+  
+      <Button
+        style="fill"
+        flexGrow={true}
+        onClick={submitButtonClicked}
+        title={submitButtonLabel}
+        bind:isDisabled={isSubmitButtonDisabled}
+      />
+    </Stack>
+  </Stack>
+</div>
 
 <script lang="ts">
+  import Button from "@components/buttons/Button.svelte"
+  import FileInput from "@components/inputs/FileInput.svelte"
+  import Stack from "@components/layout/Stack.svelte"
+  import { colors } from "@scripts/colors"
+  import { zIndexes } from "@scripts/constants"
+  import { textStyle } from "@scripts/textStyle"
   import type { DialogInputInfo } from "@shared/types/dialog"
-  import Button, { Label } from "@smui/button"
-  import Dialog, { Actions, Content, Title } from "@smui/dialog"
-  import Textfield from "@smui/textfield"
   import { isNotNullish, isNullish } from "@utils"
-
-  export let open = true
-  export let title: string | undefined
-  export let message: string | undefined
-  export let inputInfo: DialogInputInfo | undefined
-  export let hasCancelButton = false
-  export let submitButtonLabel = "Submit"
-  export let onCancel: () => void
-  export let onSubmit: ((inputValue: any) => void)
-  export let onDismiss: () => void
   
-  let textfield: Textfield
-  let submitButton: Button
+  type Props = {
+    title?: string
+    message?: string
+    inputInfo?: DialogInputInfo
+    hasCancelButton?: boolean
+    submitButtonLabel?: string
+    onCancel: () => void
+    onSubmit: ((inputValue: any) => void)
+    onDismiss: () => void
+  }
   
-  let textfieldValue = "" // Workaround for https://github.com/hperrin/svelte-material-ui/issues/628
-  let files: FileList | undefined
-  let submitted = false
+  const {
+    title,
+    message,
+    inputInfo,
+    hasCancelButton = false,
+    submitButtonLabel = "Submit",
+    onCancel,
+    onSubmit,
+    onDismiss,
+  }: Props = $props()
+  
+  let selectedFileName: string | undefined = $state()
+  let files: FileList | undefined = $state()
+  let open = $state(true)
+  let isSubmitButtonDisabled = $state(false)
   
   const selectedFile = (): File | undefined => {
     if (isNotNullish(files) && files.length !== 0) {
@@ -73,41 +139,40 @@
       if (isNotNullish(file)) {
         return new DataView(await file.arrayBuffer())
       }
-    } else {
-      if (textfieldValue !== "") {
-        return textfieldValue
-      }
     }
+  }
+  
+  const cancelButtonClicked = () => {
+    onCancel?.()
+    open = false
   }
   
   const submitButtonClicked = async () => {
     if (isNullish(inputInfo)) {
       onSubmit?.(undefined)
+      open = false
       return
     }
     
     const value = await inputValue()
     
     if (isNotNullish(value)) {
-      submitted = true
       onSubmit?.(value)
     }
+    
+    open = false
   }
   
-  $: files, textfieldValue, filesChanged()
+  $effect(() => { files; selectedFileName; filesChanged() })
   const filesChanged = async () => {
-    if (submitButton !== null && submitButton !== undefined) {
-      submitButton.getElement().disabled = isNullish(await inputValue())
+    if (isNotNullish(inputInfo)) {
+      isSubmitButtonDisabled = isNullish(await inputValue())
     }
   }
   
-  $: open, openListener()
+  $effect(() => { open; openListener() })
   const openListener = () => {
     if (!open) {
-      if (!submitted) {
-        onCancel()
-      }
-      
       onDismiss()
     }
   }
