@@ -1,4 +1,5 @@
 import { getYAML } from "@lib/utils/yamlUtils"
+import type { SettingsFromAppViewModel } from "@shared/appData/settingsFromAppViewModel"
 import { app } from "electron"
 import fs from "fs"
 import path from "path"
@@ -9,7 +10,7 @@ export const userDataPath = path.resolve(app.getPath("userData"), "userData")
 const settingsPath = path.resolve(userDataPath, "settings")
 const previousSettingsPath = path.resolve(settingsPath, ".previousSettings.yml")
 
-export const getPreviousSettings = (): any | undefined => {
+export const getPreviousSettings = (): unknown | undefined => {
   try {
     return getYAML([previousSettingsPath])
   } catch {
@@ -17,7 +18,7 @@ export const getPreviousSettings = (): any | undefined => {
   }
 }
 
-export const setPreviousSettings = (settings: any) => {
+export const setPreviousSettings = (settings: SettingsFromAppViewModel) => {
   try {
     fs.mkdirSync(settingsPath, { recursive: true })
     fs.writeFileSync(previousSettingsPath, yaml.stringify(settings))
