@@ -1,18 +1,22 @@
 import "material-icons/iconfont/material-icons.css"
 
-import MainWindow from "@components/MainWindow.svelte"
+import AppView from "@components/AppView.svelte"
 import type { MainAPI } from "@mainAPI" // Only the type is allowed to be imported, the type keyword must be placed outside of the brackets.
+import { updateColors } from "@scripts/colors"
 import { RendererAPI } from "@scripts/ipc/rendererAPI"
 import { bindMainApi, exposeWindowApi } from "electron-affinity/window"
+import { mount } from "svelte"
 
 const init = async () => {
   exposeWindowApi(new RendererAPI())
   window.mainAPI = await bindMainApi<MainAPI>("MainAPI")
   
-  new MainWindow({
+  updateColors()
+  
+  mount(AppView, {
     target: document.getElementById("mainWindow")!,
     props: {
-      initialSettings: await window.mainAPI.getPreviousSettings(),
+      initialSettings: (await window.mainAPI.getPreviousSettings()).result,
     },
   })
 }
